orchestra license headers (#5588)
