+++ Cargo.toml
@@ -83,8 +83,8 @@ members = [
 	"node/network/gossip-support",
 	"node/network/dispute-distribution",
 	"node/overseer",
-	"node/overseer/orchestra",
-	"node/overseer/orchestra/proc-macro",
+	"node/orchestra",
+	"node/orchestra/proc-macro",
 	"node/malus",
 	"node/primitives",
 	"node/service",
