+++ node/orchestra/Cargo.toml
@@ -13,7 +13,7 @@ tracing = "0.1.34"
 futures = "0.3"
 async-trait = "0.1"
 thiserror = "1"
-metered = { package = "prioritized-metered-channel", version = "0.1.1", path = "../../metered-channel" }
+metered = { package = "prioritized-metered-channel", version = "0.2.0", path = "../metered-channel" }
 orchestra-proc-macro = { version = "0.0.1", path = "./proc-macro" }
 futures-timer = "3.0.2"
 pin-project = "1.0"
