+++ node/orchestra/README.md
@@ -88,3 +88,12 @@ As always, debugging is notoriously annoying with bugged proc-macros.
 
 Therefore [`expander`](https://github.com/drahnr/expander) is employed to yield better
 error messages. Enable with `--feature=orchestra/expand`.
+
+## License
+
+Licensed under either of
+
+* Apache License, Version 2.0, (LICENSE-APACHE or <http://www.apache.org/licenses/LICENSE-2.0>)
+* MIT license (LICENSE-MIT or <http://opensource.org/licenses/MIT>)
+
+at your option.
