--- node/overseer/orchestra/proc-macro/src/parse/mod.rs
@@ -1,39 +1,0 @@
-// Copyright 2022 Parity Technologies (UK) Ltd.
-// This file is part of Polkadot.
-
-// Polkadot is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// Polkadot is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Polkadot.  If not, see <http://www.gnu.org/licenses/>.
-
-mod kw {
-	syn::custom_keyword!(event);
-	syn::custom_keyword!(signal);
-	syn::custom_keyword!(error);
-	syn::custom_keyword!(outgoing);
-	syn::custom_keyword!(gen);
-	syn::custom_keyword!(signal_capacity);
-	syn::custom_keyword!(message_capacity);
-	syn::custom_keyword!(subsystem);
-	syn::custom_keyword!(prefix);
-}
-
-mod parse_orchestra_attr;
-mod parse_orchestra_struct;
-
-mod parse_subsystem_attr;
-
-#[cfg(test)]
-mod tests;
-
-pub(crate) use self::{parse_orchestra_attr::*, parse_orchestra_struct::*};
-
-pub(crate) use self::parse_subsystem_attr::*;
