--- node/overseer/orchestra/src/tests.rs
@@ -1,39 +1,0 @@
-// Copyright 2022 Parity Technologies (UK) Ltd.
-// This file is part of Polkadot.
-
-// Polkadot is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// Polkadot is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Polkadot.  If not, see <http://www.gnu.org/licenses/>.
-
-#[test]
-#[rustversion::attr(not(stable), ignore)]
-fn ui_compile_fail() {
-	// Only run the ui tests when `RUN_UI_TESTS` is set.
-	if std::env::var("RUN_UI_TESTS").is_err() {
-		return
-	}
-
-	let t = trybuild::TestCases::new();
-	t.compile_fail("tests/ui/err-*.rs");
-}
-
-#[test]
-#[rustversion::attr(not(stable), ignore)]
-fn ui_pass() {
-	// Only run the ui tests when `RUN_UI_TESTS` is set.
-	if std::env::var("RUN_UI_TESTS").is_err() {
-		return
-	}
-
-	let t = trybuild::TestCases::new();
-	t.pass("tests/ui/ok-*.rs");
-}
